Fix incomplete abstract and title issue
In some cases the title and/or abstract obtained was incomplete
(issue #23).

This happens when the text contains html markup tags.
The most frequent ones seems to be (in descending order):
<i>, <sub>, <sup>, <b>, <mml:*>, ...?.

Example: PMID 31689885
<ArticleTitle>Gamma Irradiated <i>Rhodiola sachalinensis</i> Extract Ameliorates [...]</ArticleTitle>
Before the fix the returned title was just: 'Gamma Irradiated '
<AbstractText>The effect of <i>Rhodiola sachalinensis</i> Boriss extract irradiated [...]</ArticleTitle>
Before the fix the returned abstract was just: 'The effect of '

Fastest solution found: cleanup of tags.
It seems to fix #23 correctly, at least for non-mml tags.
NB: cleaning of nested <mml:*> tags can result in multiple blanklines.